Tests for the new design
